test rest image wip

+++ backend/rest/api.go
@@ -80,21 +80,31 @@ func (a *Api) StartApi(socket string) {
 
 	r.Use(headers)
 
-	if _, err := os.Stat(socket); err == nil {
-		err := os.Remove(socket)
+	var listener net.Listener
+	if strings.HasPrefix(socket, "tcp://") {
+		address := strings.TrimPrefix(socket, "tcp://")
+		listener, err = net.Listen("tcp", address)
 		if err != nil {
 			panic(err)
 		}
+	} else {
+		if _, err := os.Stat(socket); err == nil {
+			err := os.Remove(socket)
+			if err != nil {
+				panic(err)
+			}
+		}
+		listener, err = net.Listen("unix", socket)
+		if err != nil {
+			panic(err)
+		}
+		if err := os.Chmod(socket, 0777); err != nil {
+			log.Fatal(err)
+		}
 	}
-	unixListener, err := net.Listen("unix", socket)
-	if err != nil {
-		panic(err)
-	}
-	if err := os.Chmod(socket, 0777); err != nil {
-		log.Fatal(err)
-	}
+
 	log.Printf("Started backend (%s)\n", socket)
-	_ = http.Serve(unixListener, r)
+	_ = http.Serve(listener, r)
 
 }
